Added some sanity checks for the scripts.
- Added docker check
- Added kubectl check
- Added ibmcloud check

+++ deploy.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if ! [ -x "$(command -v kubectl)" ]; then
+  echo 'Error: kubectl is not installed.' >&2
+  exit 1
+fi
+
 kubectl -n clean delete cm/nightly-cluster-script
 kubectl -n clean create configmap nightly-cluster-script --from-file=nightly_cluster.sh
 kubectl -n clean apply -f job.yaml
